ci(nix): add rust-toolchain update script
This script can be ran to update the rust-toolchain that is used by nix
builds. Since `fenix` is a flake input end-users are able to override it
themselves to whichever version they please, however it is useful to
have the provided version up to date, particularly if the project
ends up having an MSRV.

We could likely change this script to be one that updates all flake
inputs, however I find it useful to be able to update the rust
toolchain separately from the other dependencies and have chosen to
provide a script soley for that. This matches the existing `rustup.sh`
that is used to update the rustup toolchain version. If you wish to make
this file a generic `flake.lock` updater, remove `fenix` from the call
to `nix flake update` and it will update all inputs.